Reorganise includes, modify reason-code-to-string function
A general code cleanup w.r.t. the headers and include directives globally.
Some of the includes didn't have include guards, some didn't include all their
dependencies and thus relied on the order of inclusion, and some included a ton
of headers needlessly. This has been changed to use forward declarations where
possible, and headers which define as little as possible.
Also, the function for converting the reason code to a string has been changed
to use a std::array instead of the gigantic switch statement with extraneous
break statements.
As a bonus, the compilation time is down from 10.5s to 9s, at least on my
machine.